Change the way to manage admin role
Delete setPublisherAdminAsItself, for that
the procedure that set the publisher's admin to itself should be call
only after the first member of it has been added, or there would be a
publisher that will never have a member.
Hence, the procedure should be written into the first time granting the
publisher.
Change the grantPublisher function to satisfy that.
Change the function's comment to detail the power of DEFAULT_ADMIN_ROLE
and why the publishers already be registered won't be affected by it.