Add a backend server to interact with contract
Upload a server based on django, gunicore, and nginx.
Change truffle config to upload the proxy contract on Arbitrum testnet.
Use `truffle migrate --network goerli` to upload a contract, copy the
address, and paste it to django template htmls.
Use `docker-compose up -d --build` to run a server to interact with the
contract.